refactor cv skills page refactor cv preview page
